add policy enforcement mode on top of an existing CNI plugin
This will allow Cilium to be deployed on top of existing CNI plugins
that provide networking.

For example in Flannel, a CNI plugin, it creates an L2 interface that
connects all veth's containers to it. With this CNI plugin, Cilium will
not create the cilium_host, the cilium_net nor the cilium_vxlan
interfaces as all of the networking is managed by flannel. What Cilium
will do instead is to attach a BPF program to the existing cni0
network interface.

Signed-off-by: André Martins <andre@cilium.io>