policy: return list of rules that match when determining policy enforcement
This increases performance of policy evaluation by reducing the number
of rules which are iterated over when calculating policy at each
protocol layer for an endpoint, as well as performing the matching
operation upon the endpoint selector of the rules only once, as
opposed to against all rules in the repository for each protocol layer.
Matching is an expensive operation.

Signed-off by: Ian Vernon <ian@cilium.io>