policy: add RulesSelect field to SearchContext
RulesSelect specifies whether or not to check whether a rule which is
being analyzed against its SearchContext has already been determined to select
the labels in the `To` or `From` fields within the SearchContext. This is used
to avoid using `EndpointSelector.Matches()` while analyzing rules if possible,
since said function is costly in terms of performance. Currently, this flag
is always false, so no functional change is intended here. In the future, the
list of rules that will be analyzed when generating policy for an endpoint will
only be the list of rules that select the endpoint.

Signed-off by: Ian Vernon <ian@cilium.io>